Crypto moved to own class...
